+++ yolk/src/commonMain/kotlin/ru/dimsuz/yolk/LazyKeyCache.kt
@@ -1,5 +1,7 @@
 package ru.dimsuz.yolk
 
+import kotlinx.coroutines.flow.Flow
+
 public class LazyKeyCache<K : Any, V>(
   private val keyFactory: suspend () -> K,
   expirePolicy: ExpirePolicy<K>,
@@ -7,11 +9,12 @@ public class LazyKeyCache<K : Any, V>(
   keyStore: KeyStore<K>,
   valueStore: ValueStore<K, V>,
   ticker: Ticker = Ticker.system(),
-  log: (() -> String) -> Unit = {}
 ) {
-  private val _cache = Cache(expirePolicy, fetch, keyStore, valueStore, ticker, log)
+  private val _cache = Cache(expirePolicy, fetch, keyStore, valueStore, ticker)
 
   public suspend fun load(forceRefresh: Boolean = false): V = _cache.load(keyFactory(), forceRefresh)
   public suspend fun hasExpired(): Boolean = _cache.hasExpired(keyFactory())
   public suspend fun invalidate(): Unit = _cache.invalidate(keyFactory())
+
+  public val events: Flow<CacheEvent<K>> = _cache.events
 }