+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.1.0 - 2022-01-09
+
+* Add `yolk-sqldelight-extensions` module which provides `Query<*>.asCachedFlow()`
+* Add `events` flow, rework logging to work on top of it instead of being hardcoded into Cache
+
 ## 1.0.0 - 2021-12-23
 
 * Migrate to KMM